[2019.3.4]调整录制帧率和录制动画添加

+++ app/src/main/cpp/egl/EGLCore.cpp
@@ -197,7 +197,7 @@ GLboolean EGLCore::buildVideoContext(ANativeWindow *window, EGLContext context) 
  * 设置显示时间戳pts
  * @param nsecs
  */
-void EGLCore::setPresentationTime(long long nsecs) {
+void EGLCore::setPresentationTime(uint64_t nsecs) {
     eglPresentationTimeANDROID(mDisplay,mSurface,nsecs);
 }
